Migrate Azure DevOps routes and tests to TypeScript, add tsconfig for testing, and update package dependencies
