+++ backend/package.json
@@ -27,6 +27,7 @@
     "@types/express": "^5.0.1",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.13.13",
+    "@types/supertest": "^6.0.3",
     "babel-jest": "^29.7.0",
     "eslint": "^8.57.1",
     "eslint-plugin-import": "^2.31.0",
